Move all protocol classes under common package
