+++ scim-spec/scim-spec-protocol/src/main/java/org/apache/directory/scim/spec/protocol/adapter/AttributeReferenceAdapter.java
@@ -17,7 +17,7 @@
 * under the License.
 */
 
-package org.apache.directory.scim.spec.adapter;
+package org.apache.directory.scim.spec.protocol.adapter;
 
 import jakarta.xml.bind.annotation.adapters.XmlAdapter;
