+++ scim-spec/scim-spec-protocol/src/main/java/org/apache/directory/scim/spec/protocol/adapter/FilterWrapper.java
@@ -17,7 +17,7 @@
 * under the License.
 */
 
-package org.apache.directory.scim.spec.adapter;
+package org.apache.directory.scim.spec.protocol.adapter;
 
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
