+++ scim-spec/scim-spec-protocol/src/main/java/org/apache/directory/scim/spec/protocol/search/Filter.java
@@ -26,8 +26,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
-import org.apache.directory.scim.server.filter.FilterLexer;
-import org.apache.directory.scim.server.filter.FilterParser;
+import org.apache.directory.scim.spec.protocol.filter.FilterLexer;
+import org.apache.directory.scim.spec.protocol.filter.FilterParser;
 import org.apache.directory.scim.spec.protocol.filter.ExpressionBuildingListener;
 import org.apache.directory.scim.spec.protocol.filter.FilterExpression;
 import org.apache.directory.scim.spec.protocol.filter.FilterParseException;
